fix TimeDurationPrimitiveValue of agreement time (kmk008)

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/standardtime/monthsetting/AddAgreementMonthSettingCommand.java
@@ -13,9 +13,9 @@ public class AddAgreementMonthSettingCommand {
 	
 	private String employeeId;
 
-	private Long yearMonthValue;
+	private int yearMonthValue;
 
-	private Long errorOneMonth;
+	private int errorOneMonth;
 
-	private Long alarmOneMonth;
+	private int alarmOneMonth;
 }
