+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/standardtime/primitivevalue/AlarmFourWeeks.java
@@ -12,7 +12,7 @@ import nts.arc.primitive.constraint.TimeRange;
 @TimeRange(max="672:00", min = "00:00")
 public class AlarmFourWeeks extends TimeDurationPrimitiveValue<AlarmFourWeeks> {
 	
-	public AlarmFourWeeks(Long rawValue) {
+	public AlarmFourWeeks(int rawValue) {
 		super(rawValue);
 	}
